Move testutils package to sleeper.compaction
