+++ java/compaction/compaction-core/src/test/java/sleeper/compaction/testutils/CompactionJobStatusStoreInMemoryTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.job.testutils;
+package sleeper.compaction.testutils;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
